requirements.txt
Django==3.2.25
djangorestframework==3.15.1
Requests==2.31.0
